requirements.txt
django==5.2.1
django-webpack-loader==3.2.0
social-auth-app-django==5.4.3
python-telegram-bot==22.1
python-dotenv==1.1.0
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-debug-toolbar==5.2.0
psycopg2-binary==2.9.10
celery==5.5.3
sentry-sdk==2.17.0
django-environ==0.11.2
pandas==2.2.3
ta==0.11.0
ratelimit==2.2.1
pybit==5.8.0
websocket-client==1.8.0
numpy==2.2.6